Ensure style and versions are kept up to date (#3)
When you join a team you should be able to join and write features without any stress on does the code I write looks like my teammate. Or knowing who wrote a piece of code because they use a different style. The code belongs to everyone, so it should follow a standard and a style agreed on.

Additionally, we should always try to keep our code up to date and deploy it often. Our tests should ensure things keep working

For that, we are now adding [lints](https://golangci-lint.run) and dependabot to our sample app