Add integration specs on web app (#6)
You were thinking i was going to do all this without tests? You don't know me well. I love good tests of different kinds. Here for now i just want to know that the DI stuff plays well and that i can actually call the API as if i was a user.

For that I'm using something the fx framework provides, `fxtest`. This lets me check that the dependency tree complies and it's happy and also to spin up the application to do calls to it.

Knowing that I have some test coverage I refactored the http routing to use an interface instead of relying on a struct. This will allow me have better tests for this in the future.