Use Testcontainers to spin up postgres (#10)
In this PR I change the usage of SQLite for postgres. The only reason is that while I was working on #7, the DB is stored on disk so it was annoying for my integration tests.


TestContainers is a fantastic tool that spins ups docker containers for you without much issue. This allows us to test without having a service running outside of the tests making sure that we truncate or empty every table we used. It's just nice! I've always used it in java and the GoLang version is a lot more descriptive but everything in GoLang is.

While I was working on this I wanted to learn what's the format for the DB connection which gorm.io calls DSN. This is a great post about it, https://stackoverflow.com/questions/60465376/is-there-a-standard-dsn-format

One issue that occurred as part of this change is that the e2e tests do not like to run in parallel, probably because of the use of port 8080 for the web app. I need to change this soon so that my tests run quick! Raising https://github.com/kevinrobayna/goprod/issues/11 for this