Move routes to use internal service layer instead of db (#7)
In this PR I've moved the http routes to use a public internal layer. In my opinion, this is a mandatory step so that you can change your internal stuff (i.e database/behaviour) but keep the API stable without changes.

Our API does not expose all the methods yet but this is proving a point, we can work on things internally but only expose what we need or when we need it. In my opinion the down the service layers you get the more specific the logic must be. So services should use the internal implementation of our product but only expose the external one.

Also with this PR you will see something. I've done different types of tests for different things. The repository service is being tested with a "real" app and a "real" database, as close as we can get to reality at least. This is a good practice. You don't want to rely on mocks or fakes for your database integration, you want to make sure it works as expected and then mock all your way to the resource focusing on the business logic instead of the dependencies.