Improve CI and CLI for usage
In this PR I went ahead and improved the CI and improved development with a common Makefile.

This app now provides a CLI so that we can add commands for users to use. For now, we only provide a web command but in the future if we wanted we can have one to migrate db schema only or run custom admin tasks.